修复导出时精度丢失问题

+++ ureport2-core/src/main/java/org/sbybfai/ureport/export/excel/low/Excel97Producer.java
@@ -164,7 +164,7 @@ public class Excel97Producer {
 //				        			cell.setCellType(CellType.STRING);
 				        		}else if(obj instanceof Number){
 				        			BigDecimal bigDecimal=Utils.toBigDecimal(obj);
-				        			cell.setCellValue(bigDecimal.floatValue());
+				        			cell.setCellValue(bigDecimal.doubleValue());
 //				        			cell.setCellType(CellType.NUMERIC);
 				        		}else if(obj instanceof Boolean){
 				        			cell.setCellValue((Boolean)obj);
@@ -316,7 +316,7 @@ public class Excel97Producer {
 //			        			cell.setCellType(CellType.STRING);
 			        		}else if(obj instanceof Number){
 			        			BigDecimal bigDecimal=Utils.toBigDecimal(obj);
-			        			cell.setCellValue(bigDecimal.floatValue());
+			        			cell.setCellValue(bigDecimal.doubleValue());
 //			        			cell.setCellType(CellType.NUMERIC);
 			        		}else if(obj instanceof Boolean){
 			        			cell.setCellValue((Boolean)obj);
